+++ backend/ASTU-backend-group-1/repository/user_repository_test.go
@@ -95,7 +95,7 @@ func (suite *UserRespositoryTestSuite) TestGet() {
 		}).Return(nil)
 		suite.coll.On("FindOne", mock.Anything, mock.Anything, mock.Anything).Return(singleResult)
 		defer cur.AssertExpectations(suite.T())
-		result, err := suite.userRepository.Get(domain.UserFilterOption{Username: expectedUser.Username})
+		result, err := suite.userRepository.Get(domain.UserFilterOption{Filter: domain.UserFilter{Username: expectedUser.Username}})
 		assert.NoError(err)
 		assert.Equal(expectedUser, result[0])
 	})
@@ -108,7 +108,7 @@ func (suite *UserRespositoryTestSuite) TestGet() {
 		}).Return(nil)
 		suite.coll.On("FindOne", mock.Anything, mock.Anything, mock.Anything).Return(singleResult)
 		defer cur.AssertExpectations(suite.T())
-		result, err := suite.userRepository.Get(domain.UserFilterOption{Email: expectedUser.Email})
+		result, err := suite.userRepository.Get(domain.UserFilterOption{Filter: domain.UserFilter{Email: expectedUser.Email}})
 		assert.NoError(err)
 		assert.Equal(expectedUser, result[0])
 	})
@@ -121,7 +121,7 @@ func (suite *UserRespositoryTestSuite) TestGet() {
 		}).Return(nil)
 		suite.coll.On("FindOne", mock.Anything, mock.Anything, mock.Anything).Return(singleResult)
 		defer cur.AssertExpectations(suite.T())
-		result, err := suite.userRepository.Get(domain.UserFilterOption{UserID: expectedUser.ID})
+		result, err := suite.userRepository.Get(domain.UserFilterOption{Filter: domain.UserFilter{UserId: expectedUser.ID}})
 		assert.NoError(err)
 		assert.Equal(expectedUser, result[0])
 	})
